backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "g-kentei-backend"
version = "1.0.0"
description = "G検定対策ツール - Backend API"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "G-Kentei Team" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"

dependencies = [
    # FastAPI & Web Server
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    
    # Database
    "sqlalchemy==2.0.25",
    "alembic==1.13.1",
    
    # ChromaDB & Vector Search
    "chromadb==0.4.22",
    "sentence-transformers==2.2.2",
    
    # ML & Transformers
    "torch==2.1.2",
    "transformers==4.36.2",
    "tokenizers==0.15.0",
    "numpy==1.24.4",
    
    # LLM
    "llama-cpp-python==0.2.11",
    
    # Data Processing
    "pandas==2.1.4",
    "tqdm==4.66.1",
    
    # Web Scraping
    "playwright==1.41.0",
    "beautifulsoup4==4.12.2",
    "requests==2.31.0",
    
    # Utilities
    "python-multipart==0.0.6",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    # Development & Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.23.2",
    "httpx==0.25.2",
    
    # Code Quality
    "black==23.12.1",
    "isort==5.13.2",
    "flake8==7.0.0",
    "mypy==1.8.0",
    
    # Pre-commit
    "pre-commit==3.6.0",
]

prod = [
    # Production dependencies
    "gunicorn==21.2.0",
]

[project.urls]
Homepage = "https://github.com/your-username/g-kentei"
Repository = "https://github.com/your-username/g-kentei.git"
Issues = "https://github.com/your-username/g-kentei/issues"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "llama_cpp.*",
    "playwright.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"